Big refactor,  see full message for details.
The listener and command decorators are no longer tied to an instance
of some EventsHandler, and hence are no longer methods, just functions.
Each one wraps a function (a "handler") and returns an object which
holds the handler and some metadata about the handler. These objects
are EventListener and Command.

The EventsHandler class has been removed, because listeners and
commands are only registered on a Chatbot, and no longer registered
on Plugins.

Plugins are now defined by subclassing Plugin, and the way plugins
work has been rethought. Previously a plugin was similar to a bot, in
the sense that commands and listeners would be registered on it. When a
bot was given an event to handle, it would handle any events registered
on itself, and then pass the event to any plugins added to the bot. Now,
all listeners and commands are registered directly onto a bot, and
plugins simply supply a list of Commands and EventListeners, which
are registered on a bot when that bot loads a plugin.

Bots still have method versions of the decorators, but they are merely
conveniences now instead of the main way that defining commands and
listeners work, and defined in terms of the main decorators.

When writing listeners and commands for a plugin, handlers which depend
on a plugin instance can be written by decorating plugin methods.